backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
yt-dlp==2024.3.10
transformers==4.38.2
pydub==0.25.1
moviepy==1.0.3
openai-whisper==20231117
pytube==15.0.0
torch==2.2.1
sentencepiece==0.2.0
protobuf==4.25.3
tiktoken==0.6.0
numpy==1.26.4
setuptools==69.1.1
sacremoses==0.1.1